pyproject.toml
[project]
name = "computer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.52.0",
    "Pillow>=11.0.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
exclude = ["anthropic-quickstarts"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
